mDNS Manager
Helper object that makes it easy to discover devices using mDNS on a
local network. A system can find the hostname, ip address, socket and
service name of a particular service being advertised using mDNS on a
subnet.

Defaults to tcp.http, but you can search for other services.

Events fired during the discovery process are passed back into the
client.

Also includes a small sample program that can be used to display details
of tcp http systems on the local network.